requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.1.0
python-dotenv==1.0.0
schedule==1.2.0
pywhatkit==5.4
undetected-chromedriver==3.5.4
lxml==4.9.3
fake-useragent==1.4.0 